Apply remaining local changes

+++ packages/Gametech/Marketing/src/Database/Migrations/2025_06_25_211510_add_bank_code_marketing_team_table.php
@@ -15,11 +15,14 @@ return new class extends Migration
     {
 
         Schema::table('marketing_teams', function (Blueprint $table) {
-            $table->dropColumn('bank_name'); // ลบคอลัมน์เดิม (ถ้าไม่ต้องการเก็บไว้)
-            $table->unsignedInteger('bank_code')->nullable()->after('password_hash');
 
-            // ถ้าอยากตั้ง foreign key constraint ด้วย
-            $table->foreign('bank_code')->references('code')->on('banks');
+            if (! Schema::hasColumn('marketing_teams', 'bank_code')) {
+                $table->dropColumn('bank_name'); // ลบคอลัมน์เดิม (ถ้าไม่ต้องการเก็บไว้)
+                $table->unsignedInteger('bank_code')->nullable()->after('password_hash');
+
+                // ถ้าอยากตั้ง foreign key constraint ด้วย
+                $table->foreign('bank_code')->references('code')->on('banks');
+            }
         });
 
     }
